+++ library/common/main_interface.cc
@@ -67,12 +67,19 @@ envoy_status_t reset_stream(envoy_engine_t engine, envoy_stream_t stream) {
 
 envoy_status_t set_preferred_network(envoy_engine_t engine, envoy_network_t network) {
   envoy_netconf_t configuration_key =
-      Envoy::Network::ConnectivityManager::setPreferredNetwork(network);
+      Envoy::Network::ConnectivityManagerImpl::setPreferredNetwork(network);
   Envoy::EngineHandle::runOnEngineDispatcher(engine, [configuration_key](auto& engine) -> void {
     engine.networkConnectivityManager().refreshDns(configuration_key, true);
   });
   // TODO(snowp): Should this return failure ever?
   return ENVOY_SUCCESS;
+}
+
+envoy_status_t set_proxy_settings(envoy_engine_t e, const char* host, const uint16_t port) {
+  return Envoy::EngineHandle::runOnEngineDispatcher(
+      e,
+      [proxy_settings = Envoy::Network::ProxySettings::parseHostAndPort(host, port)](auto& engine)
+          -> void { engine.networkConnectivityManager().setProxySettings(proxy_settings); });
 }
 
 envoy_status_t record_counter_inc(envoy_engine_t e, const char* elements, envoy_stats_tags tags,