+++ library/kotlin/io/envoyproxy/envoymobile/EngineBuilder.kt
@@ -1,4 +1,4 @@
-package io.envoyproxy.envoymobile
+ package io.envoyproxy.envoymobile
 
 import io.envoyproxy.envoymobile.engine.EnvoyConfiguration
 import io.envoyproxy.envoymobile.engine.EnvoyConfiguration.TrustChainVerification
@@ -36,13 +36,13 @@ open class EngineBuilder(
   protected var onEngineRunning: (() -> Unit) = {}
   protected var logger: ((String) -> Unit)? = null
   protected var eventTracker: ((Map<String, String>) -> Unit)? = null
+  protected var enableProxying = false
   private var engineType: () -> EnvoyEngine = {
     EnvoyEngineImpl(onEngineRunning, logger, eventTracker)
   }
   private var logLevel = LogLevel.INFO
   private var adminInterfaceEnabled = false
   private var grpcStatsDomain: String? = null
-  private var statsDPort: Int? = null
   private var connectTimeoutSeconds = 30
   private var dnsRefreshSeconds = 60
   private var dnsFailureRefreshSecondsBase = 2
@@ -63,7 +63,6 @@ open class EngineBuilder(
   private var h2ConnectionKeepaliveIdleIntervalMilliseconds = 1
   private var h2ConnectionKeepaliveTimeoutSeconds = 10
   private var h2ExtendKeepaliveTimeout = false
-  private var h2RawDomains = listOf<String>()
   private var maxConnectionsPerHost = 7
   private var statsFlushSeconds = 60
   private var streamIdleTimeoutSeconds = 15
@@ -76,6 +75,7 @@ open class EngineBuilder(
   private var nativeFilterChain = mutableListOf<EnvoyNativeFilterConfig>()
   private var stringAccessors = mutableMapOf<String, EnvoyStringAccessor>()
   private var keyValueStores = mutableMapOf<String, EnvoyKeyValueStore>()
+  private var statsSinks = listOf<String>()
 
   /**
    * Add a log level to use with Envoy.
@@ -90,12 +90,13 @@ open class EngineBuilder(
   }
 
   /**
-   * Add a domain to flush stats to.
-   * Passing nil disables stats emission via the gRPC stat sink.
+   * Specifies the domain (e.g. `example.com`) to use in the default gRPC stat sink to flush
+   * stats.
    *
-   * Only one of the statsd and gRPC stat sink can be enabled.
+   * Setting this value enables the gRPC stat sink, which periodically flushes stats via the gRPC
+   * MetricsService API. The flush interval is specified via addStatsFlushSeconds.
    *
-   * @param grpcStatsDomain The domain to use for stats.
+   * @param grpcStatsDomain The domain to use for the gRPC stats sink.
    *
    * @return this builder.
    */
@@ -105,17 +106,16 @@ open class EngineBuilder(
   }
 
   /**
-   * Add a loopback port to emit statsD stats to.
-   * Passing nil disables stats emission via the statsD stat sink.
+   * Adds additional stats sinks, in the form of the raw YAML/JSON configuration.
+   * Sinks added in this fashion will be included in addition to the gRPC stats sink
+   * that may be enabled via addGrpcStatsDomain.
    *
-   * Only one of the statsD and gRPC stat sink can be enabled.
-   *
-   * @param port The port to send statsD UDP packets to via loopback
+   * @param statsSinks Configurations of stat sinks to add.
    *
    * @return this builder.
    */
-  fun addStatsDPort(port: Int): EngineBuilder {
-    this.statsDPort = port
+  fun addStatsSinks(statsSinks: List<String>): EngineBuilder {
+    this.statsSinks = statsSinks
     return this
   }
 
@@ -327,6 +327,24 @@ open class EngineBuilder(
   }
 
   /**
+   * Specify whether system proxy settings should be respected. If yes, Envoy Mobile will
+   * use Android APIs to query Android Proxy settings configured on a device and will
+   * respect these settings when establishing connections with remote services.
+   *
+   * The method is introduced for experimentation purposes and as a safety guard against
+   * critical issues in the implementation of the proxying feature. It's intended to be removed
+   * after it's confirmed that proxies on Android work as expected.
+   *
+   * @param enableProxying whether to enable Envoy's support for proxies.
+   *
+   * @return This builder.
+   */
+  fun enableProxying(enableProxying: Boolean): EngineBuilder {
+    this.enableProxying = enableProxying
+    return this
+  }
+
+  /**
    * Add a rate at which to ping h2 connections on new stream creation if the connection has
    * sat idle. Defaults to 1 millisecond which effectively enables h2 ping functionality
    * and results in a connection ping on every new stream creation. Set it to
@@ -362,18 +380,6 @@ open class EngineBuilder(
    */
   fun h2ExtendKeepaliveTimeout(h2ExtendKeepaliveTimeout: Boolean): EngineBuilder {
     this.h2ExtendKeepaliveTimeout = h2ExtendKeepaliveTimeout
-    return this
-  }
-
-  /**
-   * Add a list of domains to which h2 connections will be established without protocol negotiation.
-   *
-   * @param h2RawDomains list of domains to which connections should be raw h2.
-   *
-   * @return this builder.
-   */
-  fun addH2RawDomains(h2RawDomains: List<String>): EngineBuilder {
-    this.h2RawDomains = h2RawDomains
     return this
   }
 
@@ -597,7 +603,6 @@ open class EngineBuilder(
     val engineConfiguration = EnvoyConfiguration(
       adminInterfaceEnabled,
       grpcStatsDomain,
-      statsDPort,
       connectTimeoutSeconds,
       dnsRefreshSeconds,
       dnsFailureRefreshSecondsBase,
@@ -618,7 +623,6 @@ open class EngineBuilder(
       h2ConnectionKeepaliveIdleIntervalMilliseconds,
       h2ConnectionKeepaliveTimeoutSeconds,
       h2ExtendKeepaliveTimeout,
-      h2RawDomains,
       maxConnectionsPerHost,
       statsFlushSeconds,
       streamIdleTimeoutSeconds,
@@ -630,7 +634,8 @@ open class EngineBuilder(
       nativeFilterChain,
       platformFilterChain,
       stringAccessors,
-      keyValueStores
+      keyValueStores,
+      statsSinks
     )
 
     return when (configuration) {