+++ library/objective-c/EnvoyEngine.h
@@ -365,7 +365,6 @@ extern const int kEnvoyFilterResumeStatusResumeIteration;
 @property (nonatomic, assign) UInt32 h2ConnectionKeepaliveIdleIntervalMilliseconds;
 @property (nonatomic, assign) UInt32 h2ConnectionKeepaliveTimeoutSeconds;
 @property (nonatomic, assign) BOOL h2ExtendKeepaliveTimeout;
-@property (nonatomic, strong) NSArray<NSString *> *h2RawDomains;
 @property (nonatomic, assign) UInt32 maxConnectionsPerHost;
 @property (nonatomic, assign) UInt32 statsFlushSeconds;
 @property (nonatomic, assign) UInt32 streamIdleTimeoutSeconds;
@@ -379,6 +378,7 @@ extern const int kEnvoyFilterResumeStatusResumeIteration;
 @property (nonatomic, strong) NSArray<EnvoyHTTPFilterFactory *> *httpPlatformFilterFactories;
 @property (nonatomic, strong) NSDictionary<NSString *, EnvoyStringAccessor *> *stringAccessors;
 @property (nonatomic, strong) NSDictionary<NSString *, id<EnvoyKeyValueStore>> *keyValueStores;
+@property (nonatomic, strong) NSArray<NSString *> *statsSinks;
 
 /**
  Create a new instance of the configuration.
@@ -403,7 +403,6 @@ extern const int kEnvoyFilterResumeStatusResumeIteration;
         (UInt32)h2ConnectionKeepaliveIdleIntervalMilliseconds
               h2ConnectionKeepaliveTimeoutSeconds:(UInt32)h2ConnectionKeepaliveTimeoutSeconds
                          h2ExtendKeepaliveTimeout:(BOOL)h2ExtendKeepaliveTimeout
-                                     h2RawDomains:(NSArray<NSString *> *)h2RawDomains
                             maxConnectionsPerHost:(UInt32)maxConnectionsPerHost
                                 statsFlushSeconds:(UInt32)statsFlushSeconds
                          streamIdleTimeoutSeconds:(UInt32)streamIdleTimeoutSeconds
@@ -422,7 +421,8 @@ extern const int kEnvoyFilterResumeStatusResumeIteration;
                                           stringAccessors
                                    keyValueStores:
                                        (NSDictionary<NSString *, id<EnvoyKeyValueStore>> *)
-                                           keyValueStores;
+                                           keyValueStores
+                                       statsSinks:(NSArray<NSString *> *)statsSinks;
 
 /**
  Resolves the provided configuration template using properties on this configuration.