+++ test/swift/EngineBuilderTests.swift
@@ -278,20 +278,6 @@ final class EngineBuilderTests: XCTestCase {
     self.waitForExpectations(timeout: 0.01)
   }
 
-  func testAddingH2RawDomainsAddsToConfigurationWhenRunningEnvoy() {
-    let expectation = self.expectation(description: "Run called with expected data")
-    MockEnvoyEngine.onRunWithConfig = { config, _ in
-      XCTAssertEqual(1, config.h2RawDomains.count)
-      expectation.fulfill()
-    }
-
-    _ = EngineBuilder()
-      .addEngineType(MockEnvoyEngine.self)
-      .addH2RawDomains(["h2-raw.domain"])
-      .build()
-    self.waitForExpectations(timeout: 0.01)
-  }
-
   func testSettingMaxConnectionsPerHostAddsToConfigurationWhenRunningEnvoy() {
     let expectation = self.expectation(description: "Run called with expected data")
     MockEnvoyEngine.onRunWithConfig = { config, _ in
@@ -489,7 +475,6 @@ final class EngineBuilderTests: XCTestCase {
       h2ConnectionKeepaliveIdleIntervalMilliseconds: 1,
       h2ConnectionKeepaliveTimeoutSeconds: 333,
       h2ExtendKeepaliveTimeout: true,
-      h2RawDomains: ["h2-raw.domain"],
       maxConnectionsPerHost: 100,
       statsFlushSeconds: 600,
       streamIdleTimeoutSeconds: 700,
@@ -506,7 +491,8 @@ final class EngineBuilderTests: XCTestCase {
         EnvoyHTTPFilterFactory(filterName: "TestFilter", factory: TestFilter.init),
       ],
       stringAccessors: [:],
-      keyValueStores: [:]
+      keyValueStores: [:],
+      statsSinks: []
     )
     let resolvedYAML = try XCTUnwrap(config.resolveTemplate(kMockTemplate))
     XCTAssertTrue(resolvedYAML.contains("&connect_timeout 200s"))
@@ -526,7 +512,6 @@ final class EngineBuilderTests: XCTestCase {
     XCTAssertTrue(resolvedYAML.contains("&h2_connection_keepalive_idle_interval 0.001s"))
     XCTAssertTrue(resolvedYAML.contains("&h2_connection_keepalive_timeout 333s"))
     XCTAssertTrue(resolvedYAML.contains("&h2_delay_keepalive_timeout true"))
-    XCTAssertTrue(resolvedYAML.contains("&h2_raw_domains [\"h2-raw.domain\"]"))
 
     XCTAssertTrue(resolvedYAML.contains("&max_connections_per_host 100"))
 
@@ -577,7 +562,6 @@ final class EngineBuilderTests: XCTestCase {
       h2ConnectionKeepaliveIdleIntervalMilliseconds: 1,
       h2ConnectionKeepaliveTimeoutSeconds: 333,
       h2ExtendKeepaliveTimeout: false,
-      h2RawDomains: [],
       maxConnectionsPerHost: 100,
       statsFlushSeconds: 600,
       streamIdleTimeoutSeconds: 700,
@@ -594,7 +578,8 @@ final class EngineBuilderTests: XCTestCase {
         EnvoyHTTPFilterFactory(filterName: "TestFilter", factory: TestFilter.init),
       ],
       stringAccessors: [:],
-      keyValueStores: [:]
+      keyValueStores: [:],
+      statsSinks: []
     )
     let resolvedYAML = try XCTUnwrap(config.resolveTemplate(kMockTemplate))
     XCTAssertTrue(resolvedYAML.contains("&dns_lookup_family V4_PREFERRED"))
@@ -630,7 +615,6 @@ final class EngineBuilderTests: XCTestCase {
       h2ConnectionKeepaliveIdleIntervalMilliseconds: 222,
       h2ConnectionKeepaliveTimeoutSeconds: 333,
       h2ExtendKeepaliveTimeout: false,
-      h2RawDomains: [],
       maxConnectionsPerHost: 100,
       statsFlushSeconds: 600,
       streamIdleTimeoutSeconds: 700,
@@ -643,7 +627,8 @@ final class EngineBuilderTests: XCTestCase {
       nativeFilterChain: [],
       platformFilterChain: [],
       stringAccessors: [:],
-      keyValueStores: [:]
+      keyValueStores: [:],
+      statsSinks: []
     )
     XCTAssertNil(config.resolveTemplate("{{ missing }}"))
   }